First version of the kodi hue conroller plugin.
Added two dependencies: phue, and a modified version of discoverhue
The plugin can use uPnP to discover hue bridges on the local network and explain how to connect to the bridge. Once the connection is set up, the plugin allows you to toggle rooms on/off and apply scenes to the rooms.